rename and move directory, context.params to string
