+++ src/fetching/getEvents.ts
@@ -1,4 +1,4 @@
-export default function getEvents(orgId : string | string[]) {
+export default function getEvents(orgId : string) {
     return async () : Promise<{ id: number, title: string }[]> => {
         const cRes = await fetch(`http://localhost:3000/api/orgs/${orgId}/campaigns`);
         const cData = await cRes.json();
