+++ src/fetching/getOrg.ts
@@ -1,4 +1,4 @@
-export default function getOrg(orgId : string | string[]) {
+export default function getOrg(orgId : string) {
     return async () : Promise<{ title: string }> => {
         const oRes = await fetch(`http://localhost:3000/api/orgs/${orgId}`);
         const oData = await oRes.json();
